requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.9
